requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
langgraph>=0.1.0
